fix relationships, save initial games properly and more than one

+++ freecell-backend/src/main/java/com/github/br_dr3/freecell/gateway/dto/NewGameRequestDTO.java
@@ -12,5 +12,5 @@ public class NewGameRequestDTO {
     @JsonProperty("id")
     @Min(1)
     @Max(30000)
-    private Integer seed;
+    private Long seed;
 }
